+++ src/input.rs
@@ -14,6 +14,10 @@ pub enum InputEvent {
     NextUndoneTask,
     FirstUndone,
     LastTask,
+    NewTask,
+    EditTask,
+    GitPush,
+    GitPull,
     Noop,
 }
 
@@ -27,11 +31,15 @@ pub fn handle_input() -> io::Result<InputEvent> {
                     KeyCode::Char('q') | KeyCode::Esc => return Ok(InputEvent::Quit),
                     KeyCode::Char('d') => return Ok(InputEvent::MarkDone),
                     KeyCode::Char('u') => return Ok(InputEvent::MarkUndone),
-                    KeyCode::Char('n') | KeyCode::Right => return Ok(InputEvent::NextTask),
-                    KeyCode::Char('p') | KeyCode::Left => return Ok(InputEvent::PreviousTask),
+                    KeyCode::Char('j') | KeyCode::Right => return Ok(InputEvent::NextTask),
+                    KeyCode::Char('k') | KeyCode::Left => return Ok(InputEvent::PreviousTask),
                     KeyCode::Char('N') => return Ok(InputEvent::NextUndoneTask),
                     KeyCode::Char('f') => return Ok(InputEvent::FirstUndone),
                     KeyCode::Char('l') => return Ok(InputEvent::LastTask),
+                    KeyCode::Char('n') => return Ok(InputEvent::NewTask),
+                    KeyCode::Char('e') => return Ok(InputEvent::EditTask),
+                    KeyCode::Char('P') => return Ok(InputEvent::GitPush),
+                    KeyCode::Char('p') => return Ok(InputEvent::GitPull),
                     _ => {}
                 }
             }